+++ Cargo.toml
@@ -15,6 +15,3 @@ byteorder = "1.3"
 named-binary-tag = "0.3"
 bitvec = "0.17"
 log = "0.4.11"
-
-[dev-dependencies]
-tempfile = "3.1"
